+++ src/certi/communication/messages/DdmUnassociateRegion.java
@@ -19,28 +19,13 @@
 // ----------------------------------------------------------------------------
 package certi.communication.messages;
 
+
 import certi.communication.CertiException;
-import certi.communication.MessageBuffer;
-import certi.communication.CertiMessageType;
-import certi.communication.CertiMessage;
-import certi.rti.impl.CertiHandleValuePairCollection;
-import hla.rti.AttributeHandleSet;
-import hla.rti.SuppliedAttributes;
-import certi.rti.impl.CertiExtent;
-import java.util.List;
-import hla.rti.Region;
-import hla.rti.FederateHandleSet;
-import hla.rti.SuppliedParameters;
-import certi.rti.impl.CertiLogicalTime;
-import certi.rti.impl.CertiLogicalTimeInterval;
-import hla.rti.LogicalTime;
-import hla.rti.LogicalTimeInterval;
-import hla.rti.ReflectedAttributes;
-import hla.rti.ReceivedInteraction;
+import certi.communication.*;
 
 public class DdmUnassociateRegion extends CertiMessage {
-   private long object;
-   private long region;
+   private int object;
+   private int region;
 
    public DdmUnassociateRegion() {
       super(CertiMessageType.DDM_UNASSOCIATE_REGION);
@@ -58,8 +43,8 @@ public class DdmUnassociateRegion extends CertiMessage {
    public void readMessage(MessageBuffer messageBuffer) throws CertiException {
       super.readMessage(messageBuffer); //Header 
 
-      object = messageBuffer.readLong();
-      region = messageBuffer.readLong();
+      object = messageBuffer.readInt();
+      region = messageBuffer.readInt();
    }
 
    @Override
@@ -67,19 +52,19 @@ public class DdmUnassociateRegion extends CertiMessage {
       return (super.toString() + ", object: " + object + ", region: " + region);
    }
 
-   public long getObject() {
+   public int getObject() {
       return object;
    }
 
-   public long getRegion() {
+   public int getRegion() {
       return region;
    }
 
-   public void setObject(long newObject) {
+   public void setObject(int newObject) {
       this.object = newObject;
    }
 
-   public void setRegion(long newRegion) {
+   public void setRegion(int newRegion) {
       this.region = newRegion;
    }
