+++ src/certi/communication/messages/DdmUnsubscribeAttributes.java
@@ -19,28 +19,13 @@
 // ----------------------------------------------------------------------------
 package certi.communication.messages;
 
+
 import certi.communication.CertiException;
-import certi.communication.MessageBuffer;
-import certi.communication.CertiMessageType;
-import certi.communication.CertiMessage;
-import certi.rti.impl.CertiHandleValuePairCollection;
-import hla.rti.AttributeHandleSet;
-import hla.rti.SuppliedAttributes;
-import certi.rti.impl.CertiExtent;
-import java.util.List;
-import hla.rti.Region;
-import hla.rti.FederateHandleSet;
-import hla.rti.SuppliedParameters;
-import certi.rti.impl.CertiLogicalTime;
-import certi.rti.impl.CertiLogicalTimeInterval;
-import hla.rti.LogicalTime;
-import hla.rti.LogicalTimeInterval;
-import hla.rti.ReflectedAttributes;
-import hla.rti.ReceivedInteraction;
+import certi.communication.*;
 
 public class DdmUnsubscribeAttributes extends CertiMessage {
-   private long objectClass;
-   private long region;
+   private int objectClass;
+   private int region;
 
    public DdmUnsubscribeAttributes() {
       super(CertiMessageType.DDM_UNSUBSCRIBE_ATTRIBUTES);
@@ -58,8 +43,8 @@ public class DdmUnsubscribeAttributes extends CertiMessage {
    public void readMessage(MessageBuffer messageBuffer) throws CertiException {
       super.readMessage(messageBuffer); //Header 
 
-      objectClass = messageBuffer.readLong();
-      region = messageBuffer.readLong();
+      objectClass = messageBuffer.readInt();
+      region = messageBuffer.readInt();
    }
 
    @Override
@@ -67,19 +52,19 @@ public class DdmUnsubscribeAttributes extends CertiMessage {
       return (super.toString() + ", objectClass: " + objectClass + ", region: " + region);
    }
 
-   public long getObjectClass() {
+   public int getObjectClass() {
       return objectClass;
    }
 
-   public long getRegion() {
+   public int getRegion() {
       return region;
    }
 
-   public void setObjectClass(long newObjectClass) {
+   public void setObjectClass(int newObjectClass) {
       this.objectClass = newObjectClass;
    }
 
-   public void setRegion(long newRegion) {
+   public void setRegion(int newRegion) {
       this.region = newRegion;
    }
