+++ src/certi/communication/messages/StartRegistrationForObjectClass.java
@@ -19,27 +19,12 @@
 // ----------------------------------------------------------------------------
 package certi.communication.messages;
 
+
 import certi.communication.CertiException;
-import certi.communication.MessageBuffer;
-import certi.communication.CertiMessageType;
-import certi.communication.CertiMessage;
-import certi.rti.impl.CertiHandleValuePairCollection;
-import hla.rti.AttributeHandleSet;
-import hla.rti.SuppliedAttributes;
-import certi.rti.impl.CertiExtent;
-import java.util.List;
-import hla.rti.Region;
-import hla.rti.FederateHandleSet;
-import hla.rti.SuppliedParameters;
-import certi.rti.impl.CertiLogicalTime;
-import certi.rti.impl.CertiLogicalTimeInterval;
-import hla.rti.LogicalTime;
-import hla.rti.LogicalTimeInterval;
-import hla.rti.ReflectedAttributes;
-import hla.rti.ReceivedInteraction;
+import certi.communication.*;
 
 public class StartRegistrationForObjectClass extends CertiMessage {
-   private long objectClass;
+   private int objectClass;
 
    public StartRegistrationForObjectClass() {
       super(CertiMessageType.START_REGISTRATION_FOR_OBJECT_CLASS);
@@ -56,7 +41,7 @@ public class StartRegistrationForObjectClass extends CertiMessage {
    public void readMessage(MessageBuffer messageBuffer) throws CertiException {
       super.readMessage(messageBuffer); //Header 
 
-      objectClass = messageBuffer.readLong();
+      objectClass = messageBuffer.readInt();
    }
 
    @Override
@@ -64,11 +49,11 @@ public class StartRegistrationForObjectClass extends CertiMessage {
       return (super.toString() + ", objectClass: " + objectClass);
    }
 
-   public long getObjectClass() {
+   public int getObjectClass() {
       return objectClass;
    }
 
-   public void setObjectClass(long newObjectClass) {
+   public void setObjectClass(int newObjectClass) {
       this.objectClass = newObjectClass;
    }
