+++ src/certi/communication/messages/UnsubscribeInteractionClass.java
@@ -19,27 +19,12 @@
 // ----------------------------------------------------------------------------
 package certi.communication.messages;
 
+
 import certi.communication.CertiException;
-import certi.communication.MessageBuffer;
-import certi.communication.CertiMessageType;
-import certi.communication.CertiMessage;
-import certi.rti.impl.CertiHandleValuePairCollection;
-import hla.rti.AttributeHandleSet;
-import hla.rti.SuppliedAttributes;
-import certi.rti.impl.CertiExtent;
-import java.util.List;
-import hla.rti.Region;
-import hla.rti.FederateHandleSet;
-import hla.rti.SuppliedParameters;
-import certi.rti.impl.CertiLogicalTime;
-import certi.rti.impl.CertiLogicalTimeInterval;
-import hla.rti.LogicalTime;
-import hla.rti.LogicalTimeInterval;
-import hla.rti.ReflectedAttributes;
-import hla.rti.ReceivedInteraction;
+import certi.communication.*;
 
 public class UnsubscribeInteractionClass extends CertiMessage {
-   private long interactionClass;
+   private int interactionClass;
 
    public UnsubscribeInteractionClass() {
       super(CertiMessageType.UNSUBSCRIBE_INTERACTION_CLASS);
@@ -56,7 +41,7 @@ public class UnsubscribeInteractionClass extends CertiMessage {
    public void readMessage(MessageBuffer messageBuffer) throws CertiException {
       super.readMessage(messageBuffer); //Header 
 
-      interactionClass = messageBuffer.readLong();
+      interactionClass = messageBuffer.readInt();
    }
 
    @Override
@@ -64,11 +49,11 @@ public class UnsubscribeInteractionClass extends CertiMessage {
       return (super.toString() + ", interactionClass: " + interactionClass);
    }
 
-   public long getInteractionClass() {
+   public int getInteractionClass() {
       return interactionClass;
    }
 
-   public void setInteractionClass(long newInteractionClass) {
+   public void setInteractionClass(int newInteractionClass) {
       this.interactionClass = newInteractionClass;
    }
